Move user module repository classes to the authentication module
