+++ modules/user/server-scala/src/main/scala/model/CertificateAuthTable.scala
