+++ modules/user/server-scala/src/main/scala/repositories/auth/FacebookAuthRepository.scala
@@ -1,9 +1,9 @@
-package repositories
+package repositories.auth
 
 import com.byteslounge.slickrepo.repository.Repository
 import javax.inject.Inject
+import model.auth.facebook.FacebookAuthTable.FacebookAuthTable
 import model.facebook.FacebookAuth
-import model.facebook.FacebookAuthTable.FacebookAuthTable
 import slick.ast.BaseTypedType
 import slick.jdbc.JdbcProfile
