+++ modules/user/server-scala/src/main/scala/repositories/auth/GoogleAuthSchemaInitializer.scala
@@ -1,9 +1,9 @@
-package repositories
+package repositories.auth
 
 import common.slick.SchemaInitializer
 import javax.inject.Inject
-import model.google.GoogleAuthTable
-import model.google.GoogleAuthTable.GoogleAuthTable
+import model.auth.google.GoogleAuthTable
+import model.auth.google.GoogleAuthTable.GoogleAuthTable
 import slick.lifted.TableQuery
 
 import scala.concurrent.ExecutionContext
